+++ kdt-spring-order/src/main/java/org/prgrms/kdt/CommandLine.java
@@ -7,7 +7,6 @@ import org.prgrms.kdt.io.InputConsole;
 import org.prgrms.kdt.io.Output;
 import org.prgrms.kdt.io.OutputConsole;
 
-import org.prgrms.kdt.voucher.service.MemoryVoucherService;
 import org.prgrms.kdt.voucher.service.VoucherService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -28,7 +27,7 @@ public class CommandLine implements Runnable {
 
     @Override
     public void run() {
-        MemoryVoucherService memoryVoucherService = applicationContext.getBean(MemoryVoucherService.class);
+        VoucherService voucherService = applicationContext.getBean(VoucherService.class);
         BlacklistService blacklistService = applicationContext.getBean(BlacklistService.class);
         blacklistService.setBlacklistRepository();
 
@@ -36,7 +35,7 @@ public class CommandLine implements Runnable {
             try { /* Voucher Application Running */
                 output.init();
                 String inputCommand = input.input("> ");
-                if(!CmdList.execute(inputCommand, input, output, memoryVoucherService, blacklistService)) break;
+                if(!CmdList.execute(inputCommand, input, output, voucherService, blacklistService)) break;
             } catch(Exception e) {
                 logger.error(e.getMessage());
                 output.printInvalidCmd(e.getMessage());
