+++ kdt-spring-order/src/main/java/org/prgrms/kdt/command/BlacklistCommand.java
@@ -4,12 +4,12 @@ import org.prgrms.kdt.blacklist.domain.Blacklist;
 import org.prgrms.kdt.blacklist.service.BlacklistService;
 import org.prgrms.kdt.io.Input;
 import org.prgrms.kdt.io.Output;
-import org.prgrms.kdt.voucher.service.MemoryVoucherService;
+import org.prgrms.kdt.voucher.service.VoucherService;
 
 import java.util.HashMap;
 
 public class BlacklistCommand implements Command{
-    public boolean execute(Input input, Output output, MemoryVoucherService memoryVoucherService, BlacklistService blacklistService) {
+    public boolean execute(Input input, Output output, VoucherService voucherService, BlacklistService blacklistService) {
         HashMap<String, Blacklist> map = blacklistService.getBlacklistRepository().blacklistHashMap;
         if(map.isEmpty()){
             output.emptyVoucherList();
