+++ kdt-spring-order/src/main/java/org/prgrms/kdt/command/CmdList.java
@@ -3,7 +3,6 @@ package org.prgrms.kdt.command;
 import org.prgrms.kdt.blacklist.service.BlacklistService;
 import org.prgrms.kdt.io.Input;
 import org.prgrms.kdt.io.Output;
-import org.prgrms.kdt.voucher.service.MemoryVoucherService;
 import org.prgrms.kdt.voucher.service.VoucherService;
 
 import java.util.Arrays;
@@ -23,13 +22,13 @@ public enum CmdList {
     }
 
     public static boolean execute(String inputCommand, Input input, Output output,
-                                  MemoryVoucherService memoryVoucherService, BlacklistService blacklistService) {
+                                  VoucherService voucherService, BlacklistService blacklistService) {
 
         CmdList cmdList = Arrays.stream(CmdList.values())
                 .filter(cmd -> cmd.command.equalsIgnoreCase(inputCommand))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("There is no '" + inputCommand + "'. "));
 
-        return cmdList.supplier.get().execute(input, output, memoryVoucherService, blacklistService);
+        return cmdList.supplier.get().execute(input, output, voucherService, blacklistService);
     }
 }
