+++ kdt-spring-order/src/main/java/org/prgrms/kdt/command/CreateCommand.java
@@ -4,7 +4,6 @@ import org.prgrms.kdt.blacklist.service.BlacklistService;
 import org.prgrms.kdt.io.Input;
 import org.prgrms.kdt.io.Output;
 import org.prgrms.kdt.voucher.domain.VoucherType;
-import org.prgrms.kdt.voucher.service.MemoryVoucherService;
 import org.prgrms.kdt.voucher.service.VoucherService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,7 +12,7 @@ public class CreateCommand implements Command {
     private static final Logger logger = LoggerFactory.getLogger(CreateCommand.class);
 
     @Override
-    public boolean execute(Input input, Output output, MemoryVoucherService memoryVoucherService, BlacklistService blacklistService) {
+    public boolean execute(Input input, Output output, VoucherService voucherService, BlacklistService blacklistService) {
 
         while(true) {
             try { // Get voucher type for new voucher.
@@ -26,7 +25,7 @@ public class CreateCommand implements Command {
                 int discountRate = Integer.parseInt(input.input("> "));
 
                 // Create and save new voucher.
-                memoryVoucherService.createVoucher(voucherType, discountRate);
+                voucherService.createVoucher(voucherType, discountRate);
                 break;
             } catch (IllegalArgumentException e) {
                 logger.error(e.getMessage());
