+++ kdt-spring-order/src/main/resources/datasource/voucher_db.csv
@@ -1,0 +1,3 @@
+bf5d2acc-d7b0-4cea-844d-2de41d36cdc3,FIXED,1000
+c809c0bf-dcae-49cb-8d36-c2a6629c7f5c,PERCENT,30
+30852677-12ba-4758-b708-3a48039a51c9,PERCENT,45
